% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Process.R
\name{Bayesian_Process}
\alias{Bayesian_Process}
\title{Bayesian Process for Cohen's d in t-test Designs}
\usage{
Bayesian_Process(d, N1, N2 = NULL)
}
\arguments{
\item{d}{Observed estimate of Cohen's d.}

\item{N1}{Group 1 Sample size.}

\item{N2}{Group 2 Sample size (if a two-sample design).}
}
\value{
Graphically details the process of bayesian estimation of effect size by focusing on
         the role of (1) prior, and (2) Likelihood Function to produce the posterior
         for Cohen's d.
}
\description{
Examplifies the Bayesian process to estimate Cohen's d in t-test Designs.
}
\examples{

 # Bayesian_Process(N1 = 50, N2 = 50, d = .4)

 # Bayesian_Process(N1 = 50, d = .4)
}
\author{
Reza Norouzian <rnorouzian@gmail.com>
}
