% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta_ID.R
\name{Beta_ID}
\alias{Beta_ID}
\title{Beta Prior Distribution Identifier}
\usage{
Beta_ID(Low, High, Cover = NULL)
}
\arguments{
\item{Low}{researcher's LOWEST plausible value for the parameter.}

\item{High}{researcher's HIGHEST plausible value for the parameter.}

\item{Cover}{researcher's suggested coverage for the Low and High values provided.}
}
\value{
Provides graphical as well as full textual description of a suitable Beta
        distribution for researcers based on their knowledge about how High or Low
        the parameter has been found in the liliter?ature. Also, helps researcher
        to revise their prior by issuing various messages.
}
\description{
Uses the subject matter researcher's knowledge to generate
a corresponding Beta prior distribution.
}
\details{
Uses non-linear minimization to provide an optimal Beta prior distribution.
}
\examples{
# Suppose a researcher needs a Beta prior for a parameter that in his/her
# view can't be less than 2\% and more than 90\%. The researcher believes
# these two limit values cover 90\% of all possible values that the parameter
# of interest to him/her can take:


 Beta_ID (Low = "2\%", High = "80\%", Cover = '90\%')



# User can also use any value that is between 0 and 1 for all arguments
# without using percentage sign for some or all arguments:


Beta_ID (Low = .02, High = .8, Cover = "90\%")


Beta_ID (Low = .02, High = .8, Cover = 90)
}
\author{
Reza Norouzian <rnorouzian@gmail.com>
}
