% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy_ID.R
\name{Cauchy_ID}
\alias{Cauchy_ID}
\title{Cauchy Prior Distribution Identifier}
\usage{
Cauchy_ID(Low, High, Cover = NULL)
}
\arguments{
\item{Low}{researcher's LOWEST plausible value for the parameter.}

\item{High}{researcher's HIGHEST plausible value for the parameter.}

\item{Cover}{researcher's suggested coverage for the Low and High values provided.}
}
\value{
Provides graphical as well as full textual description of a suitable Cauchy
        distribution for researcers based on their knowledge about how High or Low
        the parameter has been found in the literature. Also, helps researcher
        to revise their prior by issuing various messages.
}
\description{
Uses the subject matter researchers knowledge to generate
a corresponding Cauchy prior distribution.
}
\details{
Uses optimization techniques to provide graphical and textual information about
         an appropriate Cauchy prior distribution.
}
\examples{
# Suppose a researcher needs a Cauchy prior for a Cohen d effect size that in
# his/her view can't be less than -6 and more than +6. The researcher believes
# these two limit values cover 90\% of all possible values that this parameter
# can take:


 Cauchy_ID (Low = -6, High = 6, Cover = '90\%')



# User can also use any value that is between 0 and 1 for the argument
# cover without using percentage sign:



Cauchy_ID (Low = -6, High = 6, Cover = 90)

}
\author{
Reza Norouzian <rnorouzian@gmail.com>
}
