% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq_Testing.R
\name{Eq_Testing}
\alias{Eq_Testing}
\title{Bayesian Equivalence Testing for Cohen's d in t-test Designs}
\usage{
Eq_Testing(t, N1, N2, wid = "wi", dL = -.1, dU = .1)
}
\arguments{
\item{t}{The observed t-value.}

\item{N1}{Group 1 sample size.}

\item{N2}{Group 2 sample size (if a two-sample design).}

\item{wid}{Width or Scale of a Cauchy prior.}

\item{dL}{Lower limit of equivalence bound for d.}

\item{dU}{Upper limit of equivalence bound for d.}
}
\value{
Provides complete results of Bayesian Equivalence Testing using High Desity Region (HDI)
         with an "additional function" showing the result of changing the equivalence
         bounds to any different sizes.
}
\description{
Provides complete results of Bayesian Equivalence Testing with
an "additional function" showing the result of changing the equivalence
bounds to any different sizes.
}
\details{
Equivalnce Testing simply presents another alternative to absolute hypotheis testing (i.e., testing whether an effect is excatly 0).
         Specifically, we allow a bound of values which in our view are practically equivalent to zero to be our null hypothesis (i.e., a composite Null).
         However, most Bayeian point decisions are better decided upon using a "Decision-Theoritic" approch (see, Berger, 1985).
}
\examples{
# Suppose a researcher obtains a t-value of 1, from 2 groups each with 20 participants.
# The researcher picks a "wi", wide, Cauchy prior. The researcher uses
# dL of -0.1 and dU of 0.1 to see the result of equivalene testing:


 Eq_Testing(t = 1, N1 = 20, N2 = 20, dL = -.1, dU = .1 )
}
\author{
Reza Norouzian <rnorouzian@gmail.com>
}
