% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\name{Likelihood}
\alias{Likelihood}
\title{The Structure of a Likelihood Function}
\usage{
Likelihood(n)
}
\arguments{
\item{n}{Sample size.}
}
\value{
Graphically details the process of obtaining Likelihood values
         to produce a "Likelihood Function" using Simulation.
}
\description{
Details the Likelihood Function as an ingridient of a Bayesian estimation process.
}
\examples{

# Likelihood(n = 3)
}
\author{
Reza Norouzian <rnorouzian@gmail.com>
}
