% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal_ID.R
\name{Normal_ID}
\alias{Normal_ID}
\title{Normal Prior Distribution Identifier}
\usage{
Normal_ID(Low, High, Cover = NULL)
}
\arguments{
\item{Low}{researchers LOWEST value for the parameter.}

\item{High}{researchers HIGHEST value for the parameter.}

\item{Cover}{coverage for the Low and High values provided.}
}
\value{
Provides graphical as well as full textual description of a suitable Normal
        distribution for researcers based on their knowledge about how High or Low
        the parameter has been found in the literature. Also, helps researcher
        to revise their prior by issuing various messages.
}
\description{
Uses the subject matter researcher's knowledge to generate
a corresponding Normal prior distribution.
}
\details{
solves to provide graphical and textual information about
         an appropriate Normal prior distribution.
}
\examples{
# Suppose a researcher needs a Normal prior for a Cohen d effect size that in
# his/her view can't be less than -3 and more than +3. The researcher believes
# these two limit values cover 95\% of all possible values that this parameter
# can take:


 Normal_ID (Low = -3, High = 3, Cover = '95\%')



# User can also use any value that is between 0 and 1 for the argument
# cover without using percentage sign:



Normal_ID (Low = -3, High = 3, Cover = 95)
}
\author{
Reza Norouzian <rnorouzian@gmail.com>
}
