% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_eta_sq.R
\name{P_eta_sq}
\alias{P_eta_sq}
\title{Bayesian Estimation of "Partial Eta-Squared" in Fixed-Effects AN(C)OVA Designs}
\usage{
P_eta_sq(f, N, df1, df2, alpha, beta)
}
\arguments{
\item{f}{Observed F-value from a main or an interaction effect.}

\item{N}{Total design sample size.}

\item{df1}{Degress of freedom for the effect in question aka "Numerator".}

\item{df2}{Degrees of freedom "Error" aka "Denominator", "Within" or "Residual".}

\item{alpha}{shape1 parameter of a "Beta Prior".}

\item{beta}{shape2 parameter of a "Beta Prior".}
}
\value{
Provides graphical as well as full textual description of posterior distribution of
         "Partial Eta-Squared".
}
\description{
Provides complete Posterior as well as its various summary statistics for
"Partial Eta-Squared" in Fixed-Effects AN(C)OVA Designs.
}
\examples{

# P_eta_sq(f = 10, N = 60, df1 = 2, df2 = 57, alpha = 2.55, beta = 3.42)

# Note that, the "P_eta_sq(...)" when applied to ONE-WAY designs, is basically
# the same as "ETA-SQUARED" but when "P_eta_sq(...)" is applied to factorial
# designs it produces "PARTIAL ETA-SQUARED".
}
\references{
Norouzian, R. & Plonsky, L. (2017). Eta- and partial eta-squared in L2 research:
 A cautionary review and guide to more appropriate usage. Second Language Research, pp.1-15.
 doi: 10.1177/0267658316684904
}
\seealso{
Use function \code{\link{Beta_ID}} provided in this package to find an appropriate "Beta Prior" in your specific domain
of resaerch.

Norouzian R., & L. Plonsky (2017) available at: \url{http://journals.sagepub.com/doi/abs/10.1177/0267658316684904}

This supplemetary document also briefly explains when Partial Eta Squared should be used: \url{https://osf.io/7acye/}
}
\author{
Reza Norouzian <rnorouzian@gmail.com>
}
