% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Players.R
\name{Players}
\alias{Players}
\title{Normality in Nature}
\usage{
Players(Step, n.Players, n.Steps)
}
\arguments{
\item{Step}{ith step or ith coin toss.}

\item{n.Players}{Number of players on the field.}

\item{n.Steps}{Total number of tosses/steps planned to take.}
}
\value{
Provides positions and distribution of positions for any number of players on a sports field and generates
        a histogram after ith step.
}
\description{
Uses the example of players movements on a sports field to show the mechanics
of how a "Normal Density" comes about.
}
\details{
Shows the asymptotic behavior of binomial probability mass functions via simulation.
}
\examples{

# see 15 players' final positions and their distributions after taking
# 6 steps out of 16 total steps:


Players(Step = 16, n.Players = 1000, n.Steps = 16)



# see 1000 players' final positions and their distributions after taking
# 16 steps out of 16 total steps:


Players(Step = 16, n.Players = 1e4, n.Steps = 16)
}
\author{
Reza Norouzian <rnorouzian@gmail.com>
}
