% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Post_Cohen_d.R
\name{Post_Cohen_d}
\alias{Post_Cohen_d}
\title{Bayesian Estimation and Model Selection for Cohen's d in t-test Designs}
\usage{
Post_Cohen_d(t, N1, N2 = NULL, tl = 2, ttype, dexp = 1, wid = "wi")
}
\arguments{
\item{t}{Observed t value.}

\item{N1}{Group 1 sample size.}

\item{N2}{Group 2 sample size (if a two-sample design).}

\item{tl}{Type of Hypothesis.}

\item{ttype}{Type of t test.}

\item{dexp}{Expected direction of Cohens d before doing the study.}

\item{wid}{Width or Scale of a Cauchy prior.}
}
\value{
Provides graphical as well as full textual description of posterior distribution of
         effect size. Additionally, does Bayesian hypothesis testing producing Bayes Factors.
}
\description{
Provides complete Summary of Posterior, and Hypothesis Testing Results (i.e., Bayes Factors).
}
\examples{
# Suppose a researcher obtains a t value of 1, from 2 groups each with 20 participants.
# Thus, ttype is 2. The researcher picks a "wi", wide, Cauchy prior. Suppose
# researcher has no prefernce for the direction of Cohen d. Thus, tl 2.
# when hypothesis is two tailed, dexp is ignored. The researcher can use:


# Post_Cohen_d(t = .5, N1 = 20, ttype = 1)
}
\seealso{
\url{https://rezanorouzian.shinyapps.io/bayes-t/} for an
          interactive and more powerful version of this function
           which also compares the frequentist and the Bayesian results
           with each other.
}
\author{
Reza Norouzian <rnorouzian@gmail.com>
}
