% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior_Anatomy.R
\name{Prior_Anatomy}
\alias{Prior_Anatomy}
\title{Anatomy of Two Prior Probability Density Functions}
\usage{
Prior_Anatomy(type, half, width)
}
\arguments{
\item{type}{"Normal" or "Cauchy".}

\item{half}{if TRUE provides half of the chosen prior distribution.}

\item{width}{a numeric value determining the spread of the distribution.}
}
\value{
Graphically details the exact behavior of (1) a Normal prior, and
         (2) a Cauchy prior each presented in two form. First, as complete prior over
         all possible parameter values, and (b) as a halved prior over positive half of
         parameter values.
}
\description{
Using advanced graphics, visualizes the behavior of (1) a Normal prior, and
(2) a Cauchy priors each presented in two forms. First, as a complete prior over
all possible parameter values, and (b) as a halved prior over positive half of parameter
values.
}
\examples{

# Prior_Anatomy(type = "Normal", width = 1, half = F)
}
\author{
Reza Norouzian <rnorouzian@gmail.com>
}
